training-material/neic2024-qc101/notebooks/requirements.txt
iqm-client==17.5
matplotlib
pylatexenc
qiskit-algorithms
qiskit-iqm==13.6
qiskit[visualization]
